Nytt sket ode

+++ Eris/src/er/maths/Vec2.h
@@ -7,31 +7,31 @@ namespace er {
 
 	struct vec2
 	{
-		int32 x, y;
+		float x, y;
 
 		vec2();
-		vec2(int32 scalar);
-		vec2(int32 x, int32 y);
+		vec2(float scalar);
+		vec2(float x, float y);
 
 		vec2& Add(const vec2& other);
 		vec2& Subtract(const vec2& other);
 		vec2& Multiply(const vec2& other);
 		vec2& Divide(const vec2& other);
 
-		vec2& Add(int32 value);
-		vec2& Subtract(int32 value);
-		vec2& Multiply(int32 value);
-		vec2& Divide(int32 value);
+		vec2& Add(float value);
+		vec2& Subtract(float value);
+		vec2& Multiply(float value);
+		vec2& Divide(float value);
 
 		friend vec2 operator+(vec2 left, const vec2& right);
 		friend vec2 operator-(vec2 left, const vec2& right);
 		friend vec2 operator*(vec2 left, const vec2& right);
 		friend vec2 operator/(vec2 left, const vec2& right);
 
-		friend vec2 operator+(vec2 left, int32 value);
-		friend vec2 operator-(vec2 left, int32 value);
-		friend vec2 operator*(vec2 left, int32 value);
-		friend vec2 operator/(vec2 left, int32 value);
+		friend vec2 operator+(vec2 left, float value);
+		friend vec2 operator-(vec2 left, float value);
+		friend vec2 operator*(vec2 left, float value);
+		friend vec2 operator/(vec2 left, float value);
 
 		bool operator==(const vec2& other) const;
 		bool operator!=(const vec2& other) const;
@@ -41,17 +41,17 @@ namespace er {
 		vec2& operator*=(const vec2& other);
 		vec2& operator/=(const vec2& other);
 
-		vec2& operator+=(int32 value);
-		vec2& operator-=(int32 value);
-		vec2& operator*=(int32 value);
-		vec2& operator/=(int32 value);
+		vec2& operator+=(float value);
+		vec2& operator-=(float value);
+		vec2& operator*=(float value);
+		vec2& operator/=(float value);
 
 		bool operator<(const vec2& other) const;
 		bool operator<=(const vec2& other) const;
 		bool operator>(const vec2& other) const;
 		bool operator>=(const vec2& other) const;
 
-		int32 Distance(const vec2& other) const;
+		float Distance(const vec2& other) const;
 
 		std::string ToString() const;
