+++ Eris/src/er/maths/Vec3.h
@@ -7,31 +7,31 @@ namespace er {
 
 	struct vec3
 	{
-		int32 x, y, z;
+		float x, y, z;
 
 		vec3();
-		vec3(int32 scalar);
-		vec3(int32 x, int32 y, int32 z);
+		vec3(float scalar);
+		vec3(float x, float y, float z);
 
 		vec3& Add(const vec3& other);
 		vec3& Subtract(const vec3& other);
 		vec3& Multiply(const vec3& other);
 		vec3& Divide(const vec3& other);
 
-		vec3& Add(int32 value);
-		vec3& Subtract(int32 value);
-		vec3& Multiply(int32 value);
-		vec3& Divide(int32 value);
+		vec3& Add(float value);
+		vec3& Subtract(float value);
+		vec3& Multiply(float value);
+		vec3& Divide(float value);
 
 		friend vec3 operator+(vec3 left, const vec3& right);
 		friend vec3 operator-(vec3 left, const vec3& right);
 		friend vec3 operator*(vec3 left, const vec3& right);
 		friend vec3 operator/(vec3 left, const vec3& right);
 
-		friend vec3 operator+(vec3 left, int32 value);
-		friend vec3 operator-(vec3 left, int32 value);
-		friend vec3 operator*(vec3 left, int32 value);
-		friend vec3 operator/(vec3 left, int32 value);
+		friend vec3 operator+(vec3 left, float value);
+		friend vec3 operator-(vec3 left, float value);
+		friend vec3 operator*(vec3 left, float value);
+		friend vec3 operator/(vec3 left, float value);
 
 		bool operator==(const vec3& other) const;
 		bool operator!=(const vec3& other) const;
@@ -41,17 +41,17 @@ namespace er {
 		vec3& operator*=(const vec3& other);
 		vec3& operator/=(const vec3& other);
 
-		vec3& operator+=(int32 value);
-		vec3& operator-=(int32 value);
-		vec3& operator*=(int32 value);
-		vec3& operator/=(int32 value);
+		vec3& operator+=(float value);
+		vec3& operator-=(float value);
+		vec3& operator*=(float value);
+		vec3& operator/=(float value);
 
 		bool operator<(const vec3& other) const;
 		bool operator<=(const vec3& other) const;
 		bool operator>(const vec3& other) const;
 		bool operator>=(const vec3& other) const;
 
-		int32 Distance(const vec3& other) const;
+		float Distance(const vec3& other) const;
 
 		std::string ToString() const;
